add clearContainer method to VampirismBlockContainer
fix #800
