+++ src/main/java/de/teamlapen/vampirism/blocks/PotionTableBlock.java
@@ -11,7 +11,6 @@ import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.ServerPlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.inventory.container.SimpleNamedContainerProvider;
 import net.minecraft.util.ActionResultType;
 import net.minecraft.util.Hand;
 import net.minecraft.util.math.BlockPos;
@@ -70,12 +69,24 @@ public class PotionTableBlock extends VampirismBlockContainer {
         if (!worldIn.isRemote) {
             TileEntity tile = worldIn.getTileEntity(pos);
             if (tile instanceof PotionTableTileEntity) {
-                NetworkHooks.openGui((ServerPlayerEntity) player, (PotionTableTileEntity) tile, buffer -> buffer.writeBoolean(((PotionTableTileEntity) tile).isExtended()));
-                player.addStat(ModStats.interact_alchemical_cauldron);
+                if (((PotionTableTileEntity) tile).canOpen(player)) {
+                    NetworkHooks.openGui((ServerPlayerEntity) player, (PotionTableTileEntity) tile, buffer -> buffer.writeBoolean(((PotionTableTileEntity) tile).isExtended()));
+                    player.addStat(ModStats.interact_alchemical_cauldron);
+                }
             }
         }
 
         return ActionResultType.SUCCESS;
+    }
+
+    @Override
+    protected void clearContainer(BlockState state, World worldIn, BlockPos pos) {
+        TileEntity te = worldIn.getTileEntity(pos);
+        if (te instanceof PotionTableTileEntity) {
+            for (int i = 0; i < 8; ++i) {
+                this.dropItem(worldIn, pos, ((PotionTableTileEntity) te).removeStackFromSlot(i));
+            }
+        }
     }
 
     @Override
